refactor interface, adding gpu-driven
