+++ native/cocos/renderer/pipeline/custom/NativePipeline.cpp
@@ -22,45 +22,18 @@
  THE SOFTWARE.
 ****************************************************************************/
 
-#include <boost/utility/string_view_fwd.hpp>
-#include <sstream>
-#include "LayoutGraphFwd.h"
-#include "LayoutGraphGraphs.h"
-#include "LayoutGraphNames.h"
-#include "LayoutGraphTypes.h"
-#include "LayoutGraphUtils.h"
-#include "NativePipelineFwd.h"
-#include "NativePipelineTypes.h"
-#include "NativeRenderGraphUtils.h"
-#include "NativeUtils.h"
-#include "RenderCommonTypes.h"
-#include "RenderGraphGraphs.h"
-#include "RenderGraphTypes.h"
-#include "RenderInterfaceFwd.h"
-#include "RenderInterfaceTypes.h"
-#include "RenderingModule.h"
-#include "cocos/base/Macros.h"
-#include "cocos/base/Ptr.h"
-#include "cocos/base/StringUtil.h"
-#include "cocos/base/std/container/string.h"
-#include "cocos/math/Mat4.h"
-#include "cocos/renderer/gfx-base/GFXBuffer.h"
-#include "cocos/renderer/gfx-base/GFXDef-common.h"
-#include "cocos/renderer/gfx-base/GFXDescriptorSetLayout.h"
-#include "cocos/renderer/gfx-base/GFXDevice.h"
-#include "cocos/renderer/gfx-base/GFXInputAssembler.h"
-#include "cocos/renderer/gfx-base/GFXSwapchain.h"
-#include "cocos/renderer/gfx-base/states/GFXSampler.h"
-#include "cocos/renderer/pipeline/Enum.h"
-#include "cocos/renderer/pipeline/GlobalDescriptorSetManager.h"
 #include "cocos/renderer/pipeline/PipelineSceneData.h"
 #include "cocos/renderer/pipeline/PipelineStateManager.h"
-#include "cocos/renderer/pipeline/RenderPipeline.h"
+#include "cocos/renderer/pipeline/custom/LayoutGraphTypes.h"
+#include "cocos/renderer/pipeline/custom/LayoutGraphUtils.h"
+#include "cocos/renderer/pipeline/custom/NativeBuiltinUtils.h"
+#include "cocos/renderer/pipeline/custom/NativePipelineTypes.h"
+#include "cocos/renderer/pipeline/custom/NativeRenderGraphUtils.h"
+#include "cocos/renderer/pipeline/custom/RenderGraphGraphs.h"
+#include "cocos/renderer/pipeline/custom/RenderingModule.h"
+#include "cocos/renderer/pipeline/custom/details/GslUtils.h"
 #include "cocos/scene/RenderScene.h"
 #include "cocos/scene/RenderWindow.h"
-#include "details/DebugUtils.h"
-#include "details/GslUtils.h"
-
 #if CC_USE_DEBUG_RENDERER
     #include "profiler/DebugRenderer.h"
 #endif
