+++ native/cocos/renderer/pipeline/custom/NativeUtils.cpp
@@ -1,98 +1,12 @@
-#include "NativeUtils.h"
-#include "LayoutGraphGraphs.h"
-#include "NativePipelineTypes.h"
-#include "RenderGraphGraphs.h"
-#include "cocos/application/ApplicationManager.h"
-#include "cocos/core/Root.h"
-#include "details/GslUtils.h"
+#include "cocos/renderer/pipeline/custom/NativeUtils.h"
+#include "cocos/renderer/pipeline/custom/LayoutGraphGraphs.h"
+#include "cocos/renderer/pipeline/custom/NativePipelineTypes.h"
+#include "cocos/renderer/pipeline/custom/RenderGraphGraphs.h"
+#include "cocos/renderer/pipeline/custom/details/GslUtils.h"
 
 namespace cc {
+
 namespace render {
-
-void setupQuadVertexBuffer(gfx::Device &device, const Vec4 &viewport, float vbData[16]) {
-    auto minX = static_cast<float>(viewport.x);
-    auto maxX = static_cast<float>(viewport.x + viewport.z);
-    auto minY = static_cast<float>(viewport.y);
-    auto maxY = static_cast<float>(viewport.y + viewport.w);
-    if (device.getCapabilities().screenSpaceSignY > 0) {
-        std::swap(minY, maxY);
-    }
-    int n = 0;
-    vbData[n++] = -1.0F;
-    vbData[n++] = -1.0F;
-    vbData[n++] = minX; // uv
-    vbData[n++] = maxY;
-    vbData[n++] = 1.0F;
-    vbData[n++] = -1.0F;
-    vbData[n++] = maxX;
-    vbData[n++] = maxY;
-    vbData[n++] = -1.0F;
-    vbData[n++] = 1.0F;
-    vbData[n++] = minX;
-    vbData[n++] = minY;
-    vbData[n++] = 1.0F;
-    vbData[n++] = 1.0F;
-    vbData[n++] = maxX;
-    vbData[n++] = minY;
-}
-
-// NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
-void updateRasterPassConstants(uint32_t width, uint32_t height, Setter &setter) {
-    const auto &root = *Root::getInstance();
-    const auto shadingWidth = static_cast<float>(width);
-    const auto shadingHeight = static_cast<float>(height);
-    setter.setVec4(
-        "cc_time",
-        Vec4(
-            root.getCumulativeTime(),
-            root.getFrameTime(),
-            static_cast<float>(CC_CURRENT_ENGINE()->getTotalFrames()),
-            0.0F));
-
-    setter.setVec4(
-        "cc_screenSize",
-        Vec4(shadingWidth, shadingHeight, 1.0F / shadingWidth, 1.0F / shadingHeight));
-    setter.setVec4(
-        "cc_nativeSize",
-        Vec4(shadingWidth, shadingHeight, 1.0F / shadingWidth, 1.0F / shadingHeight));
-#if 0
-    const auto *debugView = root.getDebugView();
-    if (debugView) {
-        setter.setVec4(
-            "cc_debug_view_mode",
-            Vec4(static_cast<float>(debugView->getSingleMode()),
-                 debugView->isLightingWithAlbedo() ? 1.0F : 0.0F,
-                 debugView->isCsmLayerColoration() ? 1.0F : 0.0F,
-                 0.0F));
-        Vec4 debugPackVec{};
-        for (auto i = static_cast<uint32_t>(pipeline::DebugViewCompositeType::DIRECT_DIFFUSE);
-             i < static_cast<uint32_t>(pipeline::DebugViewCompositeType::MAX_BIT_COUNT); ++i) {
-            const auto idx = i % 4;
-            (&debugPackVec.x)[idx] = debugView->isCompositeModeEnabled(i) ? 1.0F : 0.0F;
-            const auto packIdx = static_cast<uint32_t>(floor(static_cast<float>(i) / 4.0F));
-            if (idx == 3) {
-                std::string name("cc_debug_view_composite_pack_");
-                name.append(std::to_string(packIdx + 1));
-                setter.setVec4(name, debugPackVec);
-            }
-        }
-    } else {
-        setter.setVec4("cc_debug_view_mode", Vec4(0.0F, 1.0F, 0.0F, 0.0F));
-        Vec4 debugPackVec{};
-        for (auto i = static_cast<uint32_t>(pipeline::DebugViewCompositeType::DIRECT_DIFFUSE);
-             i < static_cast<uint32_t>(pipeline::DebugViewCompositeType::MAX_BIT_COUNT); ++i) {
-            const auto idx = i % 4;
-            (&debugPackVec.x)[idx] = 1.0F;
-            const auto packIdx = static_cast<uint32_t>(floor(i / 4.0));
-            if (idx == 3) {
-                std::string name("cc_debug_view_composite_pack_");
-                name.append(std::to_string(packIdx + 1));
-                setter.setVec4(name, debugPackVec);
-            }
-        }
-    }
-#endif
-}
 
 namespace {
 
@@ -225,35 +139,6 @@ void setSamplerImpl(RenderData &data, const LayoutGraphData &lg, const ccstd::st
     data.samplers[nameID.value] = sampler;
 }
 
-LayoutGraphData::vertex_descriptor getSubpassOrPassID(
-    RenderGraph::vertex_descriptor vertID,
-    const RenderGraph &rg, const LayoutGraphData &lg) {
-    const auto queueID = parent(vertID, rg);
-    CC_ENSURES(queueID != RenderGraph::null_vertex());
-    const auto subpassOrPassID = parent(queueID, rg);
-    CC_ENSURES(subpassOrPassID != RenderGraph::null_vertex());
-    const auto passID = parent(subpassOrPassID, rg);
-
-    auto layoutID = LayoutGraphData::null_vertex();
-    if (passID == RenderGraph::null_vertex()) { // single render pass
-        const auto &layoutName = get(RenderGraph::LayoutTag{}, rg, subpassOrPassID);
-        CC_ENSURES(!layoutName.empty());
-        layoutID = locate(LayoutGraphData::null_vertex(), layoutName, lg);
-    } else { // render pass
-        const auto &passLayoutName = get(RenderGraph::LayoutTag{}, rg, passID);
-        CC_ENSURES(!passLayoutName.empty());
-        const auto passLayoutID = locate(LayoutGraphData::null_vertex(), passLayoutName, lg);
-        CC_ENSURES(passLayoutID != LayoutGraphData::null_vertex());
-
-        const auto &subpassLayoutName = get(RenderGraph::LayoutTag{}, rg, subpassOrPassID);
-        CC_ENSURES(!subpassLayoutName.empty());
-        const auto subpassLayoutID = locate(passLayoutID, subpassLayoutName, lg);
-        CC_ENSURES(subpassLayoutID != LayoutGraphData::null_vertex());
-        layoutID = subpassLayoutID;
-    }
-    CC_ENSURES(layoutID != LayoutGraphData::null_vertex());
-    return layoutID;
-}
-
 } // namespace render
+
 } // namespace cc
