+++ native/cocos/renderer/pipeline/custom/NativeUtils.h
@@ -1,8 +1,8 @@
 #pragma once
 #include "cocos/core/ArrayBuffer.h"
 #include "cocos/math/Vec4.h"
-#include "cocos/renderer/pipeline/custom/LayoutGraphTypes.h"
-#include "cocos/renderer/pipeline/custom/RenderGraphTypes.h"
+#include "cocos/renderer/pipeline/custom/LayoutGraphFwd.h"
+#include "cocos/renderer/pipeline/custom/RenderGraphFwd.h"
 #include "cocos/renderer/pipeline/custom/RenderInterfaceFwd.h"
 
 namespace cc {
@@ -10,9 +10,6 @@ namespace gfx {
 class Device;
 }
 namespace render {
-
-void setupQuadVertexBuffer(gfx::Device &device, const Vec4 &viewport, float vbData[16]);
-void updateRasterPassConstants(uint32_t width, uint32_t height, Setter &setter);
 
 void setMat4Impl(
     RenderData &data, const LayoutGraphData &lg, std::string_view name,
@@ -55,11 +52,6 @@ void setReadWriteBufferImpl(RenderData &data, const LayoutGraphData &lg, const c
 void setReadWriteTextureImpl(RenderData &data, const LayoutGraphData &lg, const ccstd::string &name, gfx::Texture *texture);
 
 void setSamplerImpl(RenderData &data, const LayoutGraphData &lg, const ccstd::string &name, gfx::Sampler *sampler);
-
-// Implementation
-LayoutGraphData::vertex_descriptor getSubpassOrPassID(
-    RenderGraph::vertex_descriptor vertID,
-    const RenderGraph &rg, const LayoutGraphData &lg);
 
 } // namespace render
